♻️ no need for more than one template
