+++ templates/device/src/my_device.cpp
