--- test/utils/test_impl_direct.hpp
@@ -1,25 +1,0 @@
-/*------------------------------------------------------------------------------
-Part of the MQSS Project, under the Apache License v2.0 with LLVM Exceptions.
-See https://llvm.org/LICENSE.txt for license information.
-SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-------------------------------------------------------------------------------*/
-
-/** @file
- * @brief This file is for testing whether a device implements all the required
- * functions without having an implementation of a driver.
- * @details It calls all the functions in the device interface to ensure that
- * they are implemented. During linking, when a function is not implemented this
- * will raise an error. Additionally, when executed, the tests check that no
- * function returns QDMI_ERROR_NOT_IMPLEMENTED.
- */
-
-#pragma once
-
-#include <gtest/gtest.h>
-
-class QDMIImplementationTest : public ::testing::Test {
-protected:
-  void SetUp() override;
-
-  void TearDown() override;
-};
